+++ app/Controllers/InterfaceController.hs
@@ -1,10 +1,11 @@
-putLine:: String
-putLine = "───────────────────────────────────────────────────\n"
+module Controllers.InterfaceController where
+    putLine:: String
+    putLine = "───────────────────────────────────────────────────\n"
 
-initialMenu:: String
-initialMenu = "Digite a letra correspondente à ação que você deseja executar\n" ++
-            "[E]studar\n" ++
-            "[A]dicionar cartões\n" ++
-            "[R]emover cartões\n" ++
-            "[M]odificar cartões\n" ++
-            "[C]ustomizar intervalos dos cartões"
+    initialMenu:: String
+    initialMenu = "Digite a letra correspondente à ação que você deseja executar\n" ++
+                "[E]studar\n" ++
+                "[A]dicionar cartões\n" ++
+                "[R]emover cartões\n" ++
+                "[M]odificar cartões\n" ++
+                "[C]ustomizar intervalos dos cartões"
